Make playlist searches by name more strict.
Add quotes around input playlist names when searching for them with the
Spotify Search API to reduce the likelihood of having an incorrect
playlist returned. This is a workaround for now; more implementation
needed in the future to ask the user if the correct playlist was found.

Also appease linters by following some of their suggested changes.