Use the current version of spotify_objects.
Instead of having locally-defined classes and subscripted dicts in the
code, the spotify_objects library will be used instead. This adds some
object orientation and type checking to the code structure, and also
allows for dot notation instead of dict subscripting.

The big changes in this patch are updating the commands to use the new
class definitions, but there are some small, new code updates, and a
handful of FIXMEs for later addressing too.

The requirements.txt file will need to be updated to include
spotify_objects when that repo is made public.